% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_dot.R
\name{dna_dots}
\alias{dna_dots}
\title{Maak een dot lijn patroon van DNA sequentie uit FASAT/FNA file}
\usage{
dna_dots(
  file,
  kleurA = "tomato",
  kleurT = "skyblue",
  kleurC = "khaki",
  kleurG = "orchid",
  size = 50
)
}
\arguments{
\item{file}{Character string pathway naar FASTA/FNA file e.g. "path/to/FNA"}

\item{kleurA}{Kleur voor base A. DEFAULT = "tomato"}

\item{kleurT}{Kleur voor base T. DEFAULT = "skyblue"}

\item{kleurC}{Kleur voor base C. DEFAULT = "khaki"}

\item{kleurG}{Kleur voor base G. DEFAULT = "orchid"}

\item{size}{Grootte van de stippen (cex). DEFAULT = 50}
}
\value{
Een stippen lijn plot
}
\description{
Visualiseert een DNA sequence als een stippen op een lijn waarbij elke base een unieke zelfgekozen kleur krijgt.
}
\examples{
fasta_path <- tempfile(fileext = ".fasta")
cat(">test\nAATCGCGTAAGCTTAGC\n", file = fasta_path)
dna_dots(fasta_path)

}
