% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_shape.R
\name{dna_shape}
\alias{dna_shape}
\title{Maak een DNA patroon uit FASTA/FNA file}
\usage{
dna_shape(file, kleur = "darkolivegreen3")
}
\arguments{
\item{file}{Path naar een FASTA of FNA file.}

\item{kleur}{Kleur van de plot. DEFAULT = "darkolivegreen3"}
}
\value{
Een shape plot van de DNA sequence.
}
\description{
Visualiseert een DNA sequence als shapes walk patroon.
}
\examples{
fasta_path <- tempfile(fileext = ".fasta")
cat(">test\nATGCGTACGTAGC\n", file = fasta_path)
dna_shape(fasta_path)
}
