% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_spiral.R
\name{dna_spiral}
\alias{dna_spiral}
\title{Maak een DNA spiraal patroon uit FASTA/FNA file}
\usage{
dna_spiral(
  file,
  kleurA = "pink",
  kleurT = "darkolivegreen1",
  kleurC = "cornsilk",
  kleurG = "lightblue2",
  shape = 5,
  size = 0.7
)
}
\arguments{
\item{file}{Character string pathway naar FASTA/FNA file e.g. "path/to/FNA"}

\item{kleurA}{Kleur voor base A. DEFAULT= "pink"}

\item{kleurT}{Kleur voor base T. DEFAULT= "darkolivegreen1"}

\item{kleurC}{Kleur voor base C. DEFAULT= "cornsilk"}

\item{kleurG}{Kleur voor base G. DEFAULT= "lightblue"}

\item{shape}{Pch vorm code. DEFAULT= 5}

\item{size}{Grootte vorm cex. DEFAULT= 1}
}
\value{
Plot met DNA sequence spiraal patroon.
}
\description{
Visualiseert een DNA sequence als een spiraal waarbij elke base een unieke zelfgekozen kleur krijgt.
}
\examples{
fasta_path <- tempfile(fileext = ".fasta")
cat(">test\nAATCGCGTA\n", file = fasta_path)
dna_spiral(fasta_path, kleurA = "pink", kleurT = "salmon", kleurC = "blue", kleurG = "brown",
shape= 15, size= 2)

}
