% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_tile.R
\name{dna_tile}
\alias{dna_tile}
\title{Maak een DNA tile patroon uit FASTA/FNA file}
\usage{
dna_tile(
  file,
  n = NULL,
  kleurA = "olivedrab3",
  kleurT = "darkolivegreen",
  kleurC = "cornsilk",
  kleurG = "peachpuff2"
)
}
\arguments{
\item{file}{Character string pathway naar FASTA/FNA file e.g. "path/to/FNA"}

\item{n}{Grootte van het raster (n*n). DEFAULT= n max}

\item{kleurA}{Kleur voor base A. DEFAULT= "olivedrab3"}

\item{kleurT}{Kleur voor base T. DEFAULT= "darkolivegreen"}

\item{kleurC}{Kleur voor base C. DEFAULT= "cornsilk"}

\item{kleurG}{Kleur voor base G. DEFAULT= "peachpuff2"}
}
\value{
Plot met DNA sequence tegel patroon.
}
\description{
Visualiseert een DNA sequence als een tile waarbij elke base een unieke zelfgekozen kleur krijgt.
}
\examples{
fasta_path <- tempfile(fileext = ".fasta")
cat(">test\nAATCGCGTA\n", file = fasta_path)
dna_tile(fasta_path, kleurA = "pink", kleurT = "salmon", kleurC = "blue", kleurG = "brown")

}
