% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{getPairwiseAlnScores}
\alias{getPairwiseAlnScores}
\title{getPairwiseAlnScores}
\usage{
getPairwiseAlnScores(project, force = FALSE)
}
\arguments{
\item{project}{}

\item{force}{}
}
\value{
project
}
\description{
getPairwiseAlnScores
}
