% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{getPlottingData}
\alias{getPlottingData}
\title{getPlottingData}
\usage{
getPlottingData(project, force = FALSE)
}
\arguments{
\item{project}{}

\item{force}{}
}
\value{
project
}
\description{
getPlottingData
}
