% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{initiateProject}
\alias{initiateProject}
\title{initiateProject}
\usage{
initiateProject(
  dob,
  given,
  folder = "data",
  top = 100,
  step = 10,
  width = 100,
  genome_seq = NULL,
  cores = 5
)
}
\arguments{
\item{dob}{}

\item{given}{}

\item{folder}{}

\item{top}{}

\item{step}{}

\item{width}{}

\item{genome_seq}{}

\item{cores}{}
}
\value{
project
}
\description{
initiateProject
}
