% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotArt}
\alias{plotArt}
\title{plotArt}
\usage{
plotArt(
  project,
  date = NULL,
  age = NULL,
  palette = scico::scale_fill_scico(palette = "batlow")
)
}
\arguments{
\item{project}{}

\item{date}{}

\item{age}{}

\item{palette}{}
}
\value{
project (invisible)
}
\description{
plotArt
}
