% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{plotDots}
\alias{plotDots}
\title{plotDots}
\usage{
plotDots(
  project,
  date = NULL,
  age = NULL,
  zoom = 1,
  ratio = 1,
  file = NULL,
  path = "output",
  orientation = "portrait",
  background = FALSE,
  maxsize = 49,
  dpi = 300,
  max.radius = 0.0075,
  theme.args = NULL,
  seed = NULL,
  ...
)
}
\description{
plotDots
}
