% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{plotLinks}
\alias{plotLinks}
\title{plotLinks}
\usage{
plotLinks(
  project,
  date = NULL,
  age = NULL,
  zoom = 1,
  ratio = 1,
  file = NULL,
  path = "output",
  orientation = "portrait",
  background = FALSE,
  maxsize = 49,
  dpi = 300,
  keep = 0.1,
  n.edges = 10,
  theme.args = NULL,
  seed = NULL,
  ...
)
}
\description{
plotLinks
}
