% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{sample_data}
\format{
data.frame
}
\usage{
sample_data
}
\description{
sample plotting data. Can be imported with dnart(..., data = dnaRt::sample_data)
}
\details{
sample_data
}
\keyword{datasets}
