% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_edges}
\alias{sample_edges}
\title{sample_edges}
\format{
integer vector
}
\usage{
sample_edges
}
\description{
A dataset containing edges from my personal dnart_project
}
\details{
sample_edges
}
\keyword{datasets}
