% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotArt}
\alias{plotArt}
\title{plotArt}
\usage{
plotArt(project, date = NULL, age = NULL, pdf = NULL, theme.args = NULL)
}
\arguments{
\item{project}{}

\item{date}{}

\item{age}{}

\item{pdf}{}

\item{theme.args}{}

\item{palette}{}
}
\value{
project (invisible)
}
\description{
plotArt
}
