% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotArt}
\alias{plotArt}
\title{plotArt}
\usage{
plotArt(
  project,
  date = NULL,
  age = NULL,
  palette = scale_fill_distiller(palette = "Spectral"),
  pdf = NULL,
  theme.args = NULL
)
}
\arguments{
\item{project}{}

\item{date}{}

\item{age}{}

\item{palette}{}

\item{pdf}{}

\item{theme.args}{}
}
\value{
project (invisible)
}
\description{
plotArt
}
