% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotArt}
\alias{plotArt}
\title{plotArt}
\usage{
plotArt(
  project,
  date = NULL,
  age = NULL,
  file = NULL,
  path = NULL,
  theme.args = NULL,
  zoom = 1,
  background = FALSE,
  ratio = 1,
  orientation = "portrait"
)
}
\arguments{
\item{project}{}

\item{date}{}

\item{age}{}

\item{path}{}

\item{theme.args}{}

\item{zoom}{}

\item{pdf}{}
}
\value{
project (invisible)
}
\description{
plotArt
}
