% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotArt}
\alias{plotArt}
\title{plotArt}
\usage{
plotArt(
  project,
  date = NULL,
  age = NULL,
  palette = scale_fill_gradientn(colors = c("#89B9BD", "#323348", "#A25543", "#FFD3B8",
    "#149698", "#8A151B", "#F04635")),
  pdf = NULL,
  theme.args = NULL
)
}
\arguments{
\item{project}{}

\item{date}{}

\item{age}{}

\item{palette}{}

\item{pdf}{}

\item{theme.args}{}
}
\value{
project (invisible)
}
\description{
plotArt
}
