% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{plotArt}
\alias{plotArt}
\title{plotArt}
\usage{
plotArt(project, date = NULL, age = NULL)
}
\arguments{
\item{project}{}

\item{date}{}

\item{age}{}
}
\value{
project (invisible)
}
\description{
plotArt
}
